+++ src/Data/Singletons/Single.hs
@@ -358,12 +358,12 @@ singInstD (InstDecl { id_cxt = cxt, id_name = inst_name
     sing_meth :: Name -> ALetDecRHS -> SgM [DDec]
     sing_meth name rhs = do
       mb_s_info <- dsReify (singValName name)
+      inst_kis <- mapM promoteType inst_tys
+      let mk_subst cls_tvbs = Map.fromList (zip (map extractTvbName cls_tvbs) inst_kis)
       (s_ty, tyvar_names, m_res_ki) <- case mb_s_info of
         Just (DVarI _ (DForallT cls_tvbs _cls_pred s_ty) _) -> do
-          let (sing_tvbs, _pred, _args, res_ty) = unravel s_ty
-          inst_kis <- mapM promoteType inst_tys
-          let subst = Map.fromList (zip (map extractTvbName cls_tvbs)
-                                        inst_kis)
+          let subst = mk_subst cls_tvbs
+              (sing_tvbs, _pred, _args, res_ty) = unravel s_ty
               m_res_ki = case res_ty of
                 _sing `DAppT` (_prom_func `DSigT` res_ki) -> Just (substKind subst res_ki)
                 _                                         -> Nothing
@@ -373,14 +373,12 @@ singInstD (InstDecl { id_cxt = cxt, id_name = inst_name
           mb_info <- dsReify name
           case mb_info of
             Just (DVarI _ (DForallT cls_tvbs _cls_pred inner_ty) _) -> do
-              let subst = Map.fromList (zip (map extractTvbName cls_tvbs)
-                                            inst_tys)
+              let subst = mk_subst cls_tvbs
               -- Make sure to expand through type synonyms here! Not doing so
               -- resulted in #167.
               raw_ty <- expand inner_ty
-              (s_ty, _num_args, tyvar_names, res_ki) <- singType (promoteValRhs name)
-                                                                 (substType subst raw_ty)
-              return (s_ty, tyvar_names, Just res_ki)
+              (s_ty, _num_args, tyvar_names, res_ki) <- singType (promoteValRhs name) raw_ty
+              return (substType subst s_ty, tyvar_names, Just (substKind subst res_ki))
             _ -> fail $ "Cannot find type of method " ++ show name
 
       let kind_map = maybe Map.empty (Map.singleton name) m_res_ki