+++ CHANGES.md
@@ -1,6 +1,12 @@
 Changelog for singletons project
 ================================
 
+next
+----
+* `singletons` now generates `a ~> b` instead of `TyFun a b -> Type` whenever
+  possible. This may require enabling `TypeInType` in code which did not
+  previously need it.
+
 2.4.1
 -----
 * Restore the `TyCon1`, `TyCon2`, etc. types. It turns out that the new